feat: payment service

+++ Mercatura/local/init.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS users (
 CREATE TABLE IF NOT EXISTS product (
     id SERIAL PRIMARY KEY,
     name TEXT NOT NULL,
-    price INTEGER NOT NULL,
+    price BIGINT NOT NULL,
     stock BIGINT NOT NULL,
     category TEXT NOT NULL
 )
